raster colorizer for wms

+++ datatypes/src/primitives/query_rectangle.rs
@@ -21,10 +21,11 @@ pub trait QueryAttributeSelection: Clone + Send + Sync {}
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 // TODO: custom deserializer that checks for duplicates (at least in API)
-pub struct BandSelection(Vec<usize>);
+pub struct BandSelection(Vec<u32>);
 
 impl BandSelection {
-    pub fn new(mut bands: Vec<usize>) -> Self {
+    pub fn new(mut bands: Vec<u32>) -> Self {
+        // TODO: Result
         fn has_no_duplicates<T: std::hash::Hash + std::cmp::Eq>(vec: &[T]) -> bool {
             let set: std::collections::HashSet<_> = vec.iter().collect();
             set.len() == vec.len()
@@ -41,41 +42,41 @@ impl BandSelection {
         Self(vec![0])
     }
 
-    pub fn first_n(n: usize) -> Self {
+    pub fn first_n(n: u32) -> Self {
         Self((0..n).collect())
     }
 
-    pub fn new_single(band: usize) -> Self {
+    pub fn new_single(band: u32) -> Self {
         Self(vec![band])
     }
 
-    pub fn count(&self) -> usize {
-        self.0.len()
+    pub fn count(&self) -> u32 {
+        self.0.len() as u32
     }
 
-    pub fn as_slice(&self) -> &[usize] {
+    pub fn as_slice(&self) -> &[u32] {
         &self.0
     }
 
-    pub fn as_vec(&self) -> Vec<usize> {
+    pub fn as_vec(&self) -> Vec<u32> {
         self.0.clone()
     }
 }
 
-impl From<usize> for BandSelection {
-    fn from(value: usize) -> Self {
+impl From<u32> for BandSelection {
+    fn from(value: u32) -> Self {
         Self(vec![value])
     }
 }
 
-impl From<Vec<usize>> for BandSelection {
-    fn from(value: Vec<usize>) -> Self {
+impl From<Vec<u32>> for BandSelection {
+    fn from(value: Vec<u32>) -> Self {
         Self::new(value)
     }
 }
 
-impl<const N: usize> From<[usize; N]> for BandSelection {
-    fn from(value: [usize; N]) -> Self {
+impl<const N: usize> From<[u32; N]> for BandSelection {
+    fn from(value: [u32; N]) -> Self {
         Self::new(value.to_vec())
     }
 }
