+++ operators/src/processing/interpolation/mod.rs
@@ -287,7 +287,7 @@ where
         tile_info: TileInformation,
         _query_rect: RasterQueryRectangle,
         start_time: TimeInstance,
-        band: usize,
+        band: u32,
     ) -> Result<Option<RasterQueryRectangle>> {
         // enlarge the spatial bounds in order to have the neighbor pixels for the interpolation
         let spatial_bounds = tile_info.spatial_partition();
